refactor: create folder infra/http and move app.ts and index.ts to architecture hexagonal
