+++ src/infra/http/app.ts
@@ -1,4 +1,5 @@
 import Fastify from 'fastify';
+
 import healthRoutes from '@modules/health/router/health.router';
 
 const app = Fastify();
