+++ src/infra/http/index.ts
@@ -1,4 +1,4 @@
-import app from '@src/app';
+import app from '@src/infra/http/app';
 import dotenv from 'dotenv';
 
 dotenv.config();
@@ -7,7 +7,6 @@ async function start() {
   try {
     const PORT = process.env.PORT || 3000;
     await app.listen({ port: Number(PORT) });
-    console.log(`Server is running on http://localhost:${PORT}`);
   } catch (err) {
     app.log.error(err);
     process.exit(1);
